PM: Introduce an Energy Model management framework
Several subsystems in the kernel (scheduler and/or thermal at the time
of writing) can benefit from knowing about the power consumed by CPUs.
Yet, this information can come from different sources (DT or SCMI for
example), in different formats, hence making it hard to exploit without
a standard API.

This commit attempts to solve this issue by introducing a centralized
Energy Model framework used to interface the data providers with the
client subsystems. This framework standardizes the API to register and
expose power numbers, and to access them safely from multiple locations.

The current design assumes that all CPUs in a frequency domain share the
same micro-architecture. As such, the energy model data is structured in
a per-frequency-domain fashion. Drivers aware of frequency domains
(typically, but not necessarily, CPUFreq drivers) are expected to register
data in the EM framework using the em_register_freq_domain() API. For
each frequency domain, the Energy Model includes a list of
<frequency, power, capacity> tuples for the OPPs of the domain alongside
a cpumask covering the involved CPUs. If needed in the future, this data
structure could be extended to enrich the model (with costs for idle
states for example).

Energy Model data structures are protected using refcounting (using the
kobject abstraction) to let client access them safely using em_cpu_get()
and em_cpu_put(). The framework is also designed to be hotplug-aware,
and is hence able to free unused data structures when all involved CPUs
are offline, and no references are held by external susbsystems.

Cc: "Rafael J. Wysocki" <rjw@rjwysocki.net>
Cc: Len Brown <len.brown@intel.com>
Cc: Pavel Machek <pavel@ucw.cz>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>