OPTIONAL: arch_topology: Start Energy Aware Scheduling
Energy Aware Scheduling starts when the scheduling domains are built if
the Energy Model (EM) is present. However, in the typical case of
Arm/Arm64 systems, the EM is provided after the scheduling domains are
first built at boot time, which results in EAS staying disabled.

This commit fixes this issue by re-building the scheduling domain from
the arch topology driver, once CPUfreq is up and running and when the
CPU capacities have been updated to their final value.

Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>