arch_topology: start Energy Aware Scheduling
Energy Aware Scheduling can be started either when the scheduling
domains are built (if the Energy Model is ready), or explicitly by
drivers. In the typical case of Arm/Arm64 systems, the Energy Model is
provided after the scheduling domains are built, so the latter solution
is required.

This commit attempts to start EAS from the arch topology driver, once
CPUfreq is up and running and when the capacity of the CPUs have been
updated to their final value.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>