gem5: Enable cpu voting to be fed to the gem5 energy ctrl.
In the MPMM investigation context it is required for the kernel
to provide the SCP code a measure of importance of the workload
executing on the cores of a given core group. This commit enables
this distinction to be made. Currently when a frequency change is
triggerd at the same time the CPU_ID, DOMAIN_ID and CPU_VOTE
memory mapped registers are written to as well.