sched/topology: Reference the platform Energy Model when available
In order to use EAS, the task scheduler has to know about the Energy
Model (EM) of the platform. This commit extends the scheduler topology
code to take references on the frequency domains objects of the EM
framework for all online CPUs. Hence, the availabilty of the EM for
those CPUs is guaranteed to the scheduler at runtime without further
checks in latency sensitive code paths (i.e. task wake-up).

A (RCU-protected) private list of online frequency domains is maintained
by the scheduler to simplify iterations. Furthermore, the availability
of an Energy Model is notified to the rest of the scheduler with a
static key to avoid impacts on non-EAS systems.

TO CONTINUE

Signed-off-by: Quentin Perret <quentin.perret@arm.com>