sched/topology: Reference the Energy Model of CPUs when available
In order to use EAS, the task scheduler has to know about the Energy
Model (EM) of the platform. This commit extends the scheduler topology
code to take references on the frequency domains objects of the EM
framework for all online CPUs, and introduces a new sched_domain
shortcut marking the lowest level at which EAS should be used.

A (RCU-protected) private list of online frequency domains is maintained
by the scheduler to enable fast iterations. Furthermore, the availability
of an EM is notified to the rest of the scheduler with a static key,
which ensures a low impact on non-EAS systems.

Generally speaking, the largest opportunity to save energy via scheduling
comes from a smarter exploitation of heterogeneous platforms (e.g.
big.LITTLE). There are cases (at least theoretical ones) where Energy
Aware Scheduling could help on symmetric systems, but measuring those
benefits in real use-cases isn't easy. Most of the time, it is a good
idea from an energy standpoint to simply spread the tasks and to keep the
OPPs as low as possible on symmetric systems, which is already what CFS
does. Although it is possible to find counter examples where this
strategy fails to save energy, the expected savings of a smarter
approach still need to be measured and further investigated.

On asymmetric systems, it is pretty clear that there is an immediate and
massive potential for saving energy with a different task placement
strategy. Hence, this patch enables EAS only for asymmetric systems, as
an incremental step. It will be possible to extend the scope of EAS to
symmetric systems later, as a second step, if proven useful.

Consequently, the sd_ea shortcut on which EAS operates is wired to the
lowest scheduling domain at which the SD_ASYM_CPUCAPACITY flag is set.
For example, it is possible to apply energy-aware scheduling within a
socket on a multi-socket system, as long as each socket has an
asymmetric topology. Cross-sockets wake-up balancing will only happen
when the system is over-utilized, or this_cpu and prev_cpu are in
different sockets.

Overall, Energy Aware Scheduling can be started if and only if:
   1. all online CPUs are covered by the EM;
   2. the EM complexity is low enough to keep scheduling overheads low;
   3. the platform has an asymmetric CPU capacity topology (detected by
      looking for the SD_ASYM_CPUCAPACITY flag in the sched_domain
      hierarchy, also pointed by the new sd_ea shortcut).

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>