sched/topology: Reference the Energy Model of CPUs when available
The existing scheduling domain hierarchy is defined to map to the cache
topology of the system. However, Energy Aware Scheduling (EAS) requires
more knowledge about the platform, and specifically needs to know about
the span of frequency domains, which are not always aligned with caches.
As an attempt to solve this issue, use the Energy Model (EM) of the
system to extend the scheduler topology code with a representation of
the frequency domains, alongside the scheduling domains.

Specifically, provide each root_domain with a linked list of the
frequency domains covering its CPUs. Create the lists in
partition_sched_domains(), in parallel with scheduling domain, if an
Energy Model is available, hence covering hotplug scenarios and/or
(exclusive) cpusets changes.

Each list is accessible to the rest of the scheduler using the
for_each_freq_domain(cpu, sfd) macro, which mimics the existing
topology iteration helpers (namely for_each_domain(cpu, sd)). As for
scheduling domains, the list is protected by RCU to ensure safe
concurrent updates.

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>