PM / OPP: Add power estimation function
The DT binding "dynamic-power-coefficient" represent the capacitance (C)
of a devices. This value can be used to estimate the active power of a
device at a specific OPP as P = C * V^2 * f with V and f respectively
the voltage and frequency of the OPP.

This commits introduces a helper function estimating the power of a
device based on the above equation.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>