OPTIONAL: cpufreq: dt: Register an Energy Model
********************************************************************
* This commit illustrates the usage of the newly introduced Energy *
* Model framework and isn't supposed to be merged as-is.           *
********************************************************************

The Energy Model framework provides an API to register the active power
of CPUs. This commit calls this API from the cpufreq-dt driver which can
estimate power using the P = C * V^2 * f equation where C, V, and f
respectively are the capacitance of the CPU and the voltage and frequency
of the peformance point.

The CPU capacitance is read from the "dynamic-power-coefficient" DT
binding, and the voltage and frequency values are obtained from PM_OPP.

Cc: "Rafael J. Wysocki" <rjw@rjwysocki.net>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>