MALI: utgard: Fixup mali_internal_sync_fence to match sync_file
The utgard driver is slightly out of sync w/ the 4.14 sync_file
structure. Basically the kref was removed and a flag value added
in 4.13, so this mostly worked as the structure size stayed
constant, but likely caused the wq to be misaligned.

So add changes to keep these two synchronized.

This likely is the cause for some mali issues we've seen
w/ 4.14.

Signed-off-by: John Stultz <john.stultz@linaro.org>