[BACKPORT] cpufreq: schedutil: Avoid missing updates for one-CPU policies
Commit 152db033d775 (schedutil: Allow cpufreq requests to be made
even when kthread kicked) made changes to prevent utilization updates
from being discarded during processing a previous request, but it
left a small window in which that still can happen in the one-CPU
policy case.  Namely, updates coming in after setting work_in_progress
in sugov_update_commit() and clearing it in sugov_work() will still
be dropped due to the work_in_progress check in sugov_update_single().

To close that window, rearrange the code so as to acquire the update
lock around the deferred update branch in sugov_update_single()
and drop the work_in_progress check from it.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-by: Juri Lelli <juri.lelli@redhat.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Joel Fernandes (Google) <joel@joelfernandes.org>

[
 Backport from upstream:
 commit a61dec744745 ("cpufreq: schedutil: Avoid missing updates for one-CPU policies")

 No major changes.

 We still miss DL integration in v4.14, thus the small conflicts fixed
 here, but actual code touched by this patch is the same as the original
 one.
]
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

Change-Id: I0abde985fe16fb83baabbebd927152a59dda65c6