ANDROID: cpufreq/schedutil: Select frequency using util_avg for RT
Schedutil always requests max frequency whenever a RT task is running.
Now that we have a better estimate of the utilization of RT runqueues,
let's use this information to make a less conservative decision, and
scale frequency according to the needs of the tasks.

A new sched_feature is introduced in order to choose the default
schedutil behaviour for RT tasks.

Change-Id: Ic9f01c8703d4f843addaa0d684012a422fe9f3b8
Signed-off-by: Quentin Perret <quentin.perret@arm.com>