H960: HACK: mailbox: Hi3660: Fixup for mailbox state machine
Now we only check the mailbox state machine at the beginning of request
the channel, and later we will not run state machine code for sending
message.  This results the state machine malfunction and the message
cannot be really handle from MCU side.

This patch is temporarily move the state machine operations into send
message flow.

Change-Id: Iee7c9dd3b840fb5990935b215f5eedadbb799e38
Signed-off-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: John Stultz <john.stultz@linaro.org>