thermal/devfreq_cooling: obtain power value from OPP core
Rely on dev_pm_opp_get_power() to provide power values for each OPP if
no custom function is provided to obtain dynamic power.

If a custom node exists for the power model of the device which contains
the dynamic_power_coefficient, a custom function needs to be provided to
compute the dynamic power based on this coefficient.

Change-Id: I6f7f4e16d20db2e00ee3601df238eb9aa2e48071
Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>