sched/topology: Introduce sched_energy_present static key
In order to ensure a minimal performance impact on non-energy-aware
systems, this patch introduces a static_key protecting the access to
Energy-Aware-Scheduling (EAS) code.

As such, the static key is enabled only if all the following conditions
are met for at least one root domain:
  1. all online CPUs of the root domain are covered by the Energy
     Model (EM);
  2. the complexity of the root domain's EM is low enough to keep
     scheduling overheads low;
  3. the platform has an asymmetric CPU capacity topology (detected by
     looking for the SD_ASYM_CPUCAPACITY flag in the sched_domain
     hierarchy).

The static key is accessed from within the sched_energy_enabled() function
which checks whether EAS can be used on a specific root domain or not.
As EAS cannot be enabled with CONFIG_ENERGY_MODEL=n, sched_energy_enabled()
is stubbed to 'false' to let the compiler remove the unsed EAS code by
constant propagation in this case.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>
Suggested-by: Morten Rasmussen <morten.rasmussen@arm.com>
Signed-off-by: Quentin Perret <quentin.perret@arm.com