sched/topology: Introduce sched_energy_present static key
In order to ensure a minimal performance impact on non-energy-aware
systems, introduce a static_key protecting the access to Energy-Aware
Scheduling (EAS) code.

The static key is set iff all the following conditions are met for at
least one root domain:
  1. all online CPUs of the root domain are covered by the Energy
     Model (EM);
  2. the complexity of the root domain's EM is low enough to keep
     scheduling overheads low;
  3. the platform has an asymmetric CPU capacity topology (detected by
     looking for the SD_ASYM_CPUCAPACITY flag in the sched_domain
     hierarchy).

The static key is accessed from within the rd_freq_domain() function
which returns the frequency domains of a root domain when they are
available. As EAS cannot be enabled with CONFIG_ENERGY_MODEL=n,
rd_freq_domain() is stubbed to 'NULL' to let the compiler remove the
unused EAS code by constant propagation.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Quentin Perret <quentin.perret@arm.com