PM / EM: Remove em_rescale_cpu_capacity
******* TO BE SQUASHED INTO ANOTHER PATCH *******

Capacity values in the EM framework make it faster to compute energy in
latency-sensistive code paths (task wake-up, for example). But the whole
em_rescale_cpu_capacity() thing is a pain in the ass, so let's get rid
of it alltogether.

struct em_freq_domain is extended with 'cost' which helps keeping the
overhead of energy calculation as low as possible.