sched/core: uclamp: track TG's RUNNABLE tasks
When a clamp value is updated we need to update the reference counting
for the CPU's clamp groups the task is affecting. This required to
release the reference to the clamp group representing the old clamp
value and to get a reference to the clamp group representing the new
clamp value.

This reference counting is actually required only for RUNNABLE tasks,
which are actually reference counting a CPU clamp group. We currently
find RUNNABLE tasks by walking the entire list of tasks within a task
group and skipping those not currently enqueued into a CPU.
However, at least on mobile systems, it's not unlikely the case of task
groups with many SLEEPING tasks. This means we can significantly reduce
the round-trip time of a clamp value change syscall by better tracking
the set of RUNNABLE tasks.

Let's use a per-TG list to track the RUNNABLE tasks enqueued in
any CPU of the system. Tasks are enqueue/dequeued from this list
whenever the get/put a reference to a clamp group. Such a specialized
list is then used to speed up the clamp value update syscalls by
focusing on updating just the actually RUNNABLE tasks.

Addition and removal from this list are protected by a per-TG spinlock,
which however enforces serialization of concurrent enqueue/dequeue
operations for all tasks of a TG.
NOTE: this is going to be a cross-CPUs synchronization which can be
quite expensive in case you have many tasks waking up or sleeping
concurrently on multiple CPUs.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>