sched/core: uclamp: enforce last task UCLAMP_MAX
When a util_max clamped task sleeps, we remove it's clamp constraints
from the CPU. However, the blocked utilization on that CPU can still be
higher then the clamp value enforced while its tasks was still running,
thus allow possibly unwanted CPU frequency increases.

This can happen when, for example, there is another (smaller) task
running on same frequency domain.
In these cases, when schedutil aggregates the utilization of all the
CPUs, it can still see the full blocked utilization of a previously
clamped CPU and thus decide to increase the frequency.

Let's fix this by using:

   uclamp_cpu_put()

to back annotate in the idle_task the clamp value of the last task
executed on each CPU. Once a CPUs is going to be idle, we then enforce
the most recently back annotated value as the CPU's clamp value while
that CPU is IDLE.

This is conveniently done by:

   uclamp_cpu_update()

once the last task of a CPU has been dequeued.

Thus, for all the time a CPU is IDLE, a call to uclamp_cpu() will
enforce the last back annotated value as UCLAMP_MAX value.

To the contrary, we do not do any enforcing for UCLAMP_MIN since, while
a CPU is idle, we don't want in general to enforce any minimum frequency
and we relay just on blocked load decay to smoothly reduce the
frequency.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Cc: Todd Kjos <tkjos@google.com>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: Dietmar Eggemann <dietmar.eggemann@arm.com>
Cc: Morten Rasmussen <morten.rasmussen@arm.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org

---
Changes in v2:
 - rabased on v4.18-rc4
 - new patch to improve a specific issue