sched/core: uclamp: enforce last task UCLAMP_MAX
When a util_max clamped task sleeps, we remove it's clamp constraints
from the CPU. However, if this was the last task running on that CPU the
CPU blocked utilization exceed the clamp value enforced while task was
running, we know allow that CPU bump its frequency.

This can happen if, for example, there is another task running on same
frequency domain, in which case once schedutil aggregates the
utilization of all the CPUs, we see the full blocked utilization of
previously clamped task.

Let's fix this by using:

   uclamp_cpu_put()

to back annotate in the idle_task the clamp value of the last task
executed on the CPU. Once a CPUs is going to be idle, we then enforce
the most recently back annotated value as the current CPU clamp value.
This is conveniently done by:

   uclamp_cpu_update()

once the last task of a CPU has been dequeued.

Thus, for all the time a CPU is IDLE, a call to uclamp_cpu() will
enforce the last back annotated value as UCLAMP_MAX value.

Instead, we do not do any enforcing for UCLAMP_MIN since, while a is
idle, we don't want to enforce any minimum frequency.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Cc: Joel Fernandes <joelaf@google.com>
Cc: Juri Lelli <juri.lelli@redhat.com>
Cc: Dietmar Eggemann <dietmar.eggemann@arm.com>
Cc: Morten Rasmussen <morten.rasmussen@arm.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org

---
Changes in v2:
 - new patch to improve a specific issue